requirements.txt
python-fasthtml
ghapi
markdown2
fh_frankenui @ git+https://github.com/answerdotai/monsterui
feedparser